services/astra-guard/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
boto3>=1.34.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
alembic>=1.13.0
zipstream-ng>=1.3.0
